Made sure the values and formats are loaded together
In some situations the number of rows for formats and data cold differ.
This caused that the formats got mixed up or at some cases the whole app
crashed. Fixed by loading them in one call and normalizing to make sure
that for each data there is a format and vice versa.